# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Chat-based personal finance tracker powered by OpenAI.",
        "title": "Finance Chat API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/quest-presets": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns DB-backed preset quests used by the reward assignment pool. Pass include_inactive=true to include archived presets.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quests"
                ],
                "summary": "List quest presets",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Include inactive/deleted presets",
                        "name": "include_inactive",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.QuestPreset"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a daily or weekly preset quest definition. User assignment/completion state remains stored separately.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quests"
                ],
                "summary": "Create a quest preset",
                "parameters": [
                    {
                        "description": "Quest preset",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.CreateQuestPresetRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.QuestPreset"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/quest-presets/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Archives a preset by setting is_active=false so existing user assignment history remains readable.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quests"
                ],
                "summary": "Delete a quest preset",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Quest preset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates editable preset quest metadata/rules. Existing user assignments keep their stored template key and history.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quests"
                ],
                "summary": "Update a quest preset",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Quest preset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Quest preset fields",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.UpdateQuestPresetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.QuestPreset"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/quests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the user's active quest batches with progress evaluated live from real transactions, plus game profile (level/XP) and streak. This read does not auto-assign quests; call POST /api/quests/generate to activate daily or weekly quests.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quests"
                ],
                "summary": "Get the user's current quest board",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.QuestBoard"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/quests/generate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Randomly picks active preset quests for the requested period if the user does not already have an active batch. Daily quests expire at next midnight; weekly quests expire after Sunday at next Monday 00:00.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quests"
                ],
                "summary": "Generate and activate daily or weekly quests",
                "parameters": [
                    {
                        "description": "Quest period to activate",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.GenerateQuestsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.QuestBoard"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/quests/reroll": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Picks a new random batch of quest templates. Rejected with 400 if the current batch still has an incomplete quest.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quests"
                ],
                "summary": "Request a new set of quests",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.QuestBoard"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.CreateQuestPresetRequest": {
            "type": "object",
            "required": [
                "difficulty",
                "key",
                "name",
                "period",
                "rule_type",
                "xp"
            ],
            "properties": {
                "accent": {
                    "type": "string",
                    "example": "#D9463B"
                },
                "behavior_tag": {
                    "type": "string",
                    "example": "impulse"
                },
                "category": {
                    "type": "string",
                    "example": "Food \u0026 Beverage"
                },
                "difficulty": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.QuestDifficulty"
                        }
                    ],
                    "example": "advanced"
                },
                "is_active": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string",
                    "example": "coffee_cap_weekly_250"
                },
                "logo": {
                    "type": "string",
                    "example": "☕"
                },
                "name": {
                    "type": "string",
                    "example": "Coffee discipline"
                },
                "period": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.QuestPeriod"
                        }
                    ],
                    "example": "weekly"
                },
                "rule_type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.QuestRuleType"
                        }
                    ],
                    "example": "coffee_spend_cap"
                },
                "sub_category": {
                    "type": "string",
                    "example": "Coffee"
                },
                "target": {
                    "type": "number",
                    "example": 500
                },
                "unit": {
                    "type": "string",
                    "example": "thb"
                },
                "xp": {
                    "type": "integer",
                    "example": 50
                }
            }
        },
...
        "controller.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "something went wrong"
                }
            }
        },
        "controller.GenerateQuestsRequest": {
            "type": "object",
            "required": [
                "period"
            ],
            "properties": {
                "period": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.QuestPeriod"
                        }
                    ],
                    "example": "daily"
                }
            }
        },
...
        "controller.UpdateQuestPresetRequest": {
            "type": "object",
            "properties": {
                "accent": {
                    "type": "string"
                },
                "behavior_tag": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "difficulty": {
                    "$ref": "#/definitions/model.QuestDifficulty"
                },
                "is_active": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "period": {
                    "$ref": "#/definitions/model.QuestPeriod"
                },
                "rule_type": {
                    "$ref": "#/definitions/model.QuestRuleType"
                },
                "sub_category": {
                    "type": "string"
                },
                "target": {
                    "type": "number"
                },
                "unit": {
                    "type": "string"
                },
                "xp": {
                    "type": "integer"
                }
            }
        },
...
        "model.QuestDifficulty": {
            "type": "string",
            "enum": [
                "basic",
                "advanced",
                "expert",
                "master",
                "grand_master",
                "easy",
                "medium",
                "hard"
            ],
            "x-enum-varnames": [
                "QuestBasic",
                "QuestAdvanced",
                "QuestExpert",
                "QuestMaster",
                "QuestGrandMaster",
                "QuestEasy",
                "QuestMedium",
                "QuestHard"
            ]
        },
        "model.QuestPeriod": {
            "type": "string",
            "enum": [
                "daily",
                "weekly"
            ],
            "x-enum-varnames": [
                "QuestPeriodDaily",
                "QuestPeriodWeekly"
            ]
        },
        "model.QuestPreset": {
            "type": "object",
            "properties": {
                "accent": {
                    "type": "string"
                },
                "behavior_tag": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "difficulty": {
                    "$ref": "#/definitions/model.QuestDifficulty"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "period": {
                    "$ref": "#/definitions/model.QuestPeriod"
                },
                "rule_type": {
                    "$ref": "#/definitions/model.QuestRuleType"
                },
                "sub_category": {
                    "type": "string"
                },
                "target": {
                    "type": "number"
                },
                "unit": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "xp": {
                    "type": "integer"
                }
            }
        },
        "model.QuestRuleType": {
            "type": "string",
            "enum": [
                "log_transaction_count",
                "spend_cap",
                "no_impulse",
                "no_spend_days",
                "category_spend_cap",
                "coffee_spend_cap",
                "beat_last_month_savings_rate"
            ],
            "x-enum-varnames": [
                "QuestRuleLogTransactionCount",
                "QuestRuleSpendCap",
                "QuestRuleNoImpulse",
                "QuestRuleNoSpendDays",
                "QuestRuleCategorySpendCap",
                "QuestRuleCoffeeSpendCap",
                "QuestRuleBeatLastMonthSavingsRate"
            ]
        },
...
        "service.GameProfileView": {
            "type": "object",
            "properties": {
                "level": {
                    "type": "integer"
                },
                "level_name": {
                    "type": "string"
                },
                "next_level_name": {
                    "type": "string"
                },
                "total_xp": {
                    "type": "integer"
                },
                "xp_for_next_level": {
                    "type": "integer"
                },
                "xp_into_level": {
                    "type": "integer"
                }
            }
        },
...
        "service.LevelRewardView": {
            "type": "object",
            "properties": {
                "cosmetic": {
                    "type": "string"
                },
                "feature": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "unlocked": {
                    "type": "boolean"
                },
                "xp_required": {
                    "type": "integer"
                }
            }
        },
...
        "service.QuestBoard": {
            "type": "object",
            "properties": {
                "can_generate_daily": {
                    "type": "boolean"
                },
                "can_generate_weekly": {
                    "type": "boolean"
                },
                "can_reroll": {
                    "type": "boolean"
                },
                "daily_quests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.QuestView"
                    }
                },
                "levels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.LevelRewardView"
                    }
                },
                "profile": {
                    "$ref": "#/definitions/service.GameProfileView"
                },
                "quests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.QuestView"
                    }
                },
                "rewards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.LevelRewardView"
                    }
                },
                "stats": {
                    "$ref": "#/definitions/service.QuestStats"
                },
                "streak": {
                    "$ref": "#/definitions/service.StreakView"
                },
                "weekly_quests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.QuestView"
                    }
                }
            }
        },
...
        "service.QuestView": {
            "type": "object",
            "properties": {
                "accent": {
                    "type": "string"
                },
                "assigned_at": {
                    "type": "string"
                },
                "completed_at": {
                    "type": "string"
                },
                "current": {
                    "type": "number"
                },
                "difficulty": {
                    "type": "string"
                },
                "done": {
                    "type": "boolean"
                },
                "expires_at": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "period": {
                    "type": "string"
                },
                "target": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                },
                "xp": {
                    "type": "integer"
                },
                "xp_awarded": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "LIFF ID token, e.g. \"Bearer \u003cid_token from liff.getIDToken()\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}